tailorBackend/procircle-be-api-main-main/backend/requirements.txt
aiomysql==0.2.0
aiosqlite==0.20.0
alembic==1.13.2
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asyncpg==0.29.0
bcrypt==4.1.3
billiard==4.2.0
boto3==1.34.144
botocore==1.34.144
Brotli==1.1.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
chain-logging==0.1.9
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clr-loader==0.2.6
coverage==7.6.1
cryptography==42.0.8
cssselect2==0.7.0
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
fonttools==4.53.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==8.4.0
iniconfig==2.0.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
kombu==5.3.7
lxml==5.2.2
Mako==1.3.5
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.0.0
oborpc==0.1.5
openai==1.76.2
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pdfkit==1.0.0
pillow==10.4.0
pluggy==1.5.0
prompt_toolkit==3.0.47
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydyf==0.11.0
pyee==11.1.0
Pygments==2.18.0
PyJWT==2.8.0
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
PyMySQL==1.1.1
pyphen==0.16.0
pyppeteer==2.0.0
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
python-worker==2.2.2
pythonnet==3.0.3
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.14.0
setuptools==72.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
threadpoolctl==3.5.0
tinycss2==1.3.0
tqdm==4.66.5
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==1.26.19
uvicorn==0.30.1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==10.4
wkhtmltopdf==0.2
wsproto==1.2.0
zipp==3.20.0
zopfli==0.2.3
